+++ src/main/java/com/example/datn/repository/LichSuHoaDonRepo.java
@@ -11,5 +11,5 @@ import java.util.List;
 @Repository
 public interface LichSuHoaDonRepo extends JpaRepository<LichSuHoaDon,Integer> {
     @Query("select l from LichSuHoaDon l where l.hoaDon.id=:idHoaDon")
-    List<LichSuHoaDon> findLichSuHoaDonByIdHoaDon(@Param("idHoaDon") int idHoaDon);
+    List<LichSuHoaDon> findLichSuHoaDonByIdHoaDon(@Param("idHoaDon") Long idHoaDon);
 }
